#89 Add resourcesItem template
