+++ templates/resources.jsx
@@ -1,7 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import Adapt from 'core/js/adapt';
 import a11y from 'core/js/a11y';
-import device from 'core/js/device';
 import { classes, templates } from 'core/js/reactHelpers';
 
 export default function Resources (props) {
@@ -25,20 +24,6 @@ export default function Resources (props) {
 
   function resourcesGetColumnCount(resources) {
     return _.uniq(_.pluck(resources, '_type')).length + 1;// add 1 for the 'All' button column
-  }
-
-  /**
-   * IE doesn't support the 'download' attribute
-   * https://github.com/adaptlearning/adapt_framework/issues/1559
-   * and iOS just opens links with that attribute in the same window
-   * https://github.com/adaptlearning/adapt_framework/issues/1852
-   */
-  function resourcesForceDownload(filename, _forceDownload) {
-    if (device.browser === 'internet explorer' || device.OS === 'ios') {
-      return false;
-    }
-
-    return (_forceDownload || filename);
   }
 
   const [selectedFilter, setSelectedFilter] = useState('all');
@@ -111,37 +96,16 @@ export default function Resources (props) {
         <div role="list">
 
           {resources.map(({ title, description, _link, _type, _isGlobal, filename, _forceDownload }, index) =>
-            <div className={classes([
-              'resources__item drawer__item',
-              `is-${_type}`,
-              _isGlobal && 'is-global',
-              (!['all', _type].includes(selectedFilter)) && 'u-display-none'
-            ])}
-            role="listitem"
-            key={index}>
-
-              <a href={_link} className="resources__item-btn drawer__item-btn"
-                data-type={_type}
-                download={resourcesForceDownload(filename, _forceDownload) && filename }
-                target="_blank"
-                rel="noreferrer"
-                aria-label={title}>
-
-                {title &&
-                <div className="resources__item-title drawer__item-title">
-                  <div className="resources__item-title-inner drawer__item-title-inner" dangerouslySetInnerHTML={{ __html: title }} />
-                </div>
-                }
-
-                {description &&
-                <div className="resources__item-body drawer__item-body">
-                  <div className="resources__item-body-inner drawer__item-body-inner" dangerouslySetInnerHTML={{ __html: description }} />
-                </div>
-                }
-
-              </a>
-
-            </div>
+            <templates.resourcesItem {...props}
+              key={index}
+              title={title}
+              description={description}
+              _link={_link}
+              _type={_type}
+              _isGlobal={_isGlobal}
+              selectedFilter={selectedFilter}
+              filename={ filename}
+              _forceDownload={_forceDownload} />
           )};
 
         </div>
